Backend/securefileshare/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
Django==4.2.18
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
pycparser==2.22
PyJWT==2.10.1
sqlparse==0.5.3
typing-extensions==4.12.2